requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
dotenv
alembic
passlib
bcrypt==4.0.1
python-jose